+++ dyndnsc/plugins/builtin.py
@@ -10,8 +10,8 @@ on importing this module. If importing a plugin fails, it will be ignored.
 from ..common.load import load_class as _load_plugin
 
 _BUILTINS = (
-    ("dyndnsc.plugins.notify.growl", "GrowlPlugin"),
-    ("dyndnsc.plugins.notify.osxnotify", "OSXNotifyPlugin"),
+    ("dyndnsc.plugins.notify.growl", "Growl"),
+    ("dyndnsc.plugins.notify.osxnotify", "OSXNotify"),
 )
 
 PLUGINS = {plug for plug in (_load_plugin(m, c) for m, c in _BUILTINS) if plug is not None}